Timer and mode transition
Permits timer interrupts; when a timer interrupt occurs, the ISR moves into the timer_handler, which turns OFF state mode and turns ON value mode. 

Notes: PIE does NOT need to be modified unless we're using nested interrupts. The JP1 IRQ line is disabled in the timer_handler because we don't want any interrupts to occur. Upon re-enabling the JP1 interrupts (when we're leaving the timer_handler), we clear the Edge Capture Register pre-emptively in case there was [incorrect] data there.